Fix multiple instances of tables and jobs client creation issue (#338)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

In the existing codebase for every operation task creation (every job
submission thread) a new TablesClient and JobsClient instances are
created. That results multiple such clients creation and effectively not
taking the advantage of http connection pooling. Instead default
connection pool (pool size is 500) is created for each new client
created which will effectively exhaust all the resources on the client
side. One job submission thread makes just one call to job service to
launch job so JobClient with pool size of 500 is waster of resources.
This is a major issue and was unnoticed so far due to the nature of
sequential submit which slows down over a period of time and
sequential/low connection usage. We often times get netty issues in the
job scheduler logs such a resource exhaustion, netty closing connection
etc. when we adjust parallelism or regular cadence execution, but the
actual root cause was not noticed so far.
In the new multi-mode operation model, jobs submission frequency is
higher and submission and status check happens in parallel and hence
more connection usage. So with the multi mode operation started seeing
more connection issues, more frequent DNS lookup failures after couple
of hours of execution. Eventually netty starts closing the connections
and netty runs out of resources. So the fix done here is to use single
instance of TablesClient and JobsClient with connection pooling.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.