Fix active->active replication (#344)
## Summary
For Active-> Active table replication, a new property
`openhouse.isTableReplicated` is required to identify Primary tables
which are replicated. It is used to a tablesService layer to skip
certain validations which will fail for requests coming from replication
due to differences in table properties.

This property was not getting added to table metadata and hence
replication process was failing in active-> active cases.

Fix: Add it while curating the table properties in table creation step.
Default value should be false.
Only tables which are created for replication use case should have
openhouse.isTableReplicated = true

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [x] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
Unit tests added/updated

<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.