Restore soft deleted tables endpoint (#363)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->


[Issue](https://github.com/linkedin/openhouse/issues/#nnn) Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

Add additional endpoint for restoring a soft deleted table that was
dropped.
Restore table relies on no external system deleting the original
location where the table is stored, so any cleanup job needs to be
mindful not to delete the physical locations where the table is located,
otherwise restore will create a corrupted table.

Restore table is a PUT endpoint on
`/databases/<db>/tables/<tbl_name>?deletedAtMs=<timestamp>`
It is expected that the user or system will list the timestamps to
restore from before restore is invoked.

This PR also fixes a bug on HTS side that needs to validate that the
table does not exist on user_table_row first before restoring the soft
deleted table


## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done

Tested on local docker setup
Created a table:
```
{
    "tableId": "t4",
    "databaseId": "d3",
    "baseTableVersion": "INITIAL_VERSION",
    "clusterId": "LocalHadoopCluster",
    "schema": "{\"type\": \"struct\", \"fields\": [{\"id\": 1,\"required\": true,\"name\": \"id\",\"type\": \"string\"},{\"id\": 2,\"required\": true,\"name\": \"name\",\"type\": \"string\"},{\"id\": 3,\"required\": true,\"name\": \"ts\",\"type\": \"timestamp\"}]}",
    "timePartitioning": {
        "columnName": "ts",
        "granularity": "HOUR"
    },
    "clustering": [
        {
            "columnName": "name"
        }
    ],
    "tableProperties": {
        "key": "value"
    }
}
```
Dropped the table manually from mysql (soft delete behavior) and moved
the record to `soft_deleted_user_table_row`

```
mysql> INSERT INTO soft_deleted_user_table_row (
    ->     table_id,
    ->     database_id,
    ->     deleted_at_ms,
    ->     version,
    ->     metadata_location,
    ->     storage_type,
    ->     creation_time,
    ->     purge_after_ms
    -> )
    -> SELECT
    ->     table_id,
    ->     database_id,
    ->     1723489328000,                    -- deleted_at_ms (current timestamp: 2025-08-12T16:52:08-04:00)
    ->     version,
    ->     metadata_location,
    ->     storage_type,
    ->     creation_time,
    ->     1724094128000                     -- purge_after_ms (7 days later: 2025-08-19T16:52:08-04:00)
    -> FROM user_table_row
    -> ;
```

Restore the table 
```
curl "${curlArgs[@]}" -XPOST http://localhost:8000/v1/databases/d3/tables/t4/restore\?deletedAtMs\=1723489328000
```

Validated the failure cases:
1. When soft deleted table doesn't exist:
```
{
    "status": "NOT_FOUND",
    "error": "Not Found",
    "message": "Soft deleted user table not found with timestamp 17551305860001",
    "stacktrace": null,
    "cause": ""
}
```

2. When table already exists (due to recreate):
```
{
    "status": "CONFLICT",
    "error": "Conflict",
    "message": "Table d3.t4 already exists",
    "stacktrace": null,
    "cause": "Not Available"
}
```

<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.