Update the openhouse dependency to use iceberg build of linkedin fork. (#377)
## Summary

Update to use the linkedin fork of iceberg 1.2.0 version for spark-3.1
module.


To backport data locality and other critical patches from iceberg
1.5/spark3.5 into iceberg.1.2/spark-3.1 module, we created a fork of
apache iceberg into LinkedIn iceberg.

As a first step in integration, we are updating the open-house to use
the iceberg build published from the linkedin-iceberg fork.


## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [x] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [x] No tests added or updated. Since this is a maven coordinate switch
from open source iceberg to linkedin iceberg fork.
- [x] Some other form of testing like staging or soak time in
production.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.

---------

Co-authored-by: Shanthoosh Pazhanjur Venkataraman <svenkataraman+LinkedIn@linkedin.com>