Ensure interval start time is consistent when defining replication policy (#380)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn)] Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

Replication policy is defined using the following syntax:
```
spark.sql("ALTER TABLE openhouse.<tableName> SET POLICY (REPLICATION=({destination:'<destCluster>', interval:<replicationInterval>}, {destination:'<destCluster2>', interval:<replicationInterval>}))")

```
where replication interval can 3h, 6h, 12h, 1d, 2d, etc.

There is an issue when converting this policy to a Quartz Cron schedule,
where the start time may be inconsistent with what is intended in the
policy for the hourly granularities.

For example, one observed output is:
```
0 0 17/3 ? * * *
```
which is every 3 hours starting at 5pm each day. Which is unintended as
it will only run for 5pm, 8pm, and 11pm each day.

To fix this, ensure that the start time is always less than the
interval. For example if the granularity is 3 hours, the start time
should only be from 12am, 1am, 2am, 3am.

We could also change it so that the start time is `*` which sets the
start time to whenever the replication policy is defined and runs
consistently, but this may run into a thundering herd situation for
automation.


## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.
Unit tests

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.