Repairing metrics (#393)
## Summary
in production instances, this database cardinality combined with many
histogram buckets per database adds
  1. catalog_metadata_retrieval_latency_seconds_bucket: 19.1MB (59%)
  2. catalog_metadata_update_latency_seconds_bucket:     5.2MB (16%)
  3. authentication_token_auth_success_total:            3.5MB (11%)
  4. list_acl_policies_latency_seconds_bucket:           2.4MB  (7%)
  5. http_server_requests_seconds_bucket:                0.9MB  (3%)
                                                        ------
     Top 5 metrics:                                    ~31.1MB (97%)

but we are capped at 20MB per scrape using our internal metrics platform

solution here is to live without our database dimension 😢 

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [X] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [X] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.